Move to ApiTestCase
